rename files and functions
