+++ installer/setup.py
@@ -55,15 +55,15 @@ exts = [
               sources=[pjoin('Utilities', 'akima.c')],
               include_dirs=[pjoin(numpy.get_include(), 'numpy')],
               extra_compile_args=[]),
-    Extension('wind.fwind',
-              sources=[pjoin('wind', 'fwind.f90')],
+    Extension('wind._windmodels',
+              sources=[pjoin('wind', 'windmodels.f90')],
               include_dirs=[pjoin(numpy.get_include(), 'numpy')],
               extra_compile_args=['-g']),
-    Extension('Utilities.fmaputils',
+    Extension('Utilities._maputils',
                   sources=[pjoin('Utilities', 'maputils.f90')],
                   include_dirs=[pjoin(numpy.get_include(), 'numpy')],
                   extra_compile_args=['-g']),
-    Extension('PressureInterface.fpressureProfile',
+    Extension('PressureInterface._pressureProfile',
                   sources=[pjoin('PressureInterface', 'pressureProfile.f90')],
                   include_dirs=[pjoin(numpy.get_include(), 'numpy')],
                   extra_compile_args=['-g'])
